refactor synapse into its own class and three instances of ABCs
