+++ rdb_type/mysql.py
@@ -1,6 +1,5 @@
 """MYSQL
 """
-from typing import List, Dict
 import pandas as pd
 import numpy as np
 import sqlalchemy as sa
@@ -58,21 +57,59 @@ class MySQL(RDBType):
         self.engine = engine2
         self.metadata = sa.MetaData()
 
-    def get_table_id_from_name(self, table_name: str) -> str:
-        return table_name
-
-    def get_table_name_from_id(self, table_id: str) -> str:
-        return table_id
+    def execute_sql_query(self, query: str) -> pd.DataFrame:
+        result = self.execute_sql_statement(query).fetchall()
+        table = pd.DataFrame(result)
+        return table
 
     def execute_sql_statement(self, statement: str):
         with self.engine.connect().execution_options(autocommit=True) as conn:
             result = conn.execute(statement)
         return result
 
-    def execute_sql_query(self, query: str) -> pd.DataFrame:
-        result = self.execute_sql_statement(query).fetchall()
-        table = pd.DataFrame(result)
-        return table
+    def update_table(self, data: pd.DataFrame, table_config: DBObjectConfig):
+        table_names = self.get_table_names()
+        table_name = table_config.name
+        if table_name not in table_names:
+            self.add_table(table_name, table_config)
+        self.upsert_table_rows(table_name, data)
+
+    def get_table_names(self) -> list[str]:
+        inspector = sa.inspect(self.engine)
+        return inspector.get_table_names()
+
+    def add_table(self, table_name: str, table_config: DBObjectConfig):
+        columns = self._create_columns(table_config)
+        sa.Table(table_name, self.metadata, *columns)
+        self.metadata.create_all(self.engine)
+
+    def upsert_table_rows(self, table_name: str, data: pd.DataFrame):
+        data = data.replace({np.nan: None})
+        rows = data.to_dict("records")
+        table = sa.Table(table_name, self.metadata, autoload_with=self.engine)
+        for row in rows:
+            statement = insert(table).values(row).on_duplicate_key_update(**row)
+            with self.engine.connect().execution_options(autocommit=True) as conn:
+                conn.execute(statement)
+
+    def drop_table(self, table_name: str):
+        self.execute_sql_statement(f"DROP TABLE IF EXISTS {table_name};")
+        self.metadata.clear()
+
+    def delete_table_rows(
+        self, table_name: str, data: pd.DataFrame, table_config: DBObjectConfig
+    ):
+        primary_keys = table_config.primary_keys
+        for col in primary_keys:
+            if col not in list(data.columns):
+                raise ValueError(f"primary key: {col} missing from data")
+        data = data[primary_keys]
+        tuples = list(data.itertuples(index=False, name=None))
+        tuples = [(f"'{i}'" for i in tup) for tup in tuples]
+        tuple_strings = ["(" + ",".join(tup) + ")" for tup in tuples]
+        tuple_string = ",".join(tuple_strings)
+        statement = f"DELETE FROM {table_name} WHERE ({','.join(primary_keys)}) IN ({tuple_string})"
+        self.execute_sql_statement(statement)
 
     def query_table(
         self, table_name: str, table_config: DBObjectConfig
@@ -85,12 +122,7 @@ class MySQL(RDBType):
                 table = table.astype({att.name: pandas_value})
         return table
 
-    def add_table(self, table_name: str, table_config: DBObjectConfig):
-        columns = self._create_columns(table_config)
-        sa.Table(table_name, self.metadata, *columns)
-        self.metadata.create_all(self.engine)
-
-    def _create_columns(self, table_config: DBObjectConfig) -> List[sa.Column]:
+    def _create_columns(self, table_config: DBObjectConfig) -> list[sa.Column]:
         columns = []
         for att in table_config.attributes:
             att_name = att.name
@@ -113,101 +145,3 @@ class MySQL(RDBType):
         if table_config.primary_keys != []:
             columns.append(sa.PrimaryKeyConstraint(*table_config.primary_keys))
         return columns
-
-    def drop_table(self, table_name: str):
-        self.execute_sql_statement(f"DROP TABLE IF EXISTS {table_name};")
-        self.metadata.clear()
-
-    def add_table_column(self, table_name: str, column_name: str, datatype: str):
-        self.execute_sql_statement(
-            f"ALTER TABLE {table_name} ADD {column_name} {datatype};"
-        )
-
-    def drop_table_column(self, table_name: str, column_name: str):
-        self.execute_sql_statement(f"ALTER TABLE {table_name} DROP {column_name};")
-
-    def delete_table_rows(
-        self, table_name: str, data: pd.DataFrame, table_config: DBObjectConfig
-    ):
-        primary_keys = table_config.primary_keys
-        for col in primary_keys:
-            if col not in list(data.columns):
-                raise ValueError(f"primary key: {col} missing from data")
-        data = data[primary_keys]
-        tuples = list(data.itertuples(index=False, name=None))
-        tuples = [(f"'{i}'" for i in tup) for tup in tuples]
-        tuple_strings = ["(" + ",".join(tup) + ")" for tup in tuples]
-        tuple_string = ",".join(tuple_strings)
-        statement = f"DELETE FROM {table_name} WHERE ({','.join(primary_keys)}) IN ({tuple_string})"
-        self.execute_sql_statement(statement)
-
-    def upsert_table_rows(self, table_name: str, data: pd.DataFrame):
-        data = data.replace({np.nan: None})
-        rows = data.to_dict("records")
-        table = sa.Table(table_name, self.metadata, autoload_with=self.engine)
-        for row in rows:
-            statement = insert(table).values(row).on_duplicate_key_update(**row)
-            with self.engine.connect().execution_options(autocommit=True) as conn:
-                conn.execute(statement)
-
-    def get_table_names(self) -> List[str]:
-        inspector = sa.inspect(self.engine)
-        return inspector.get_table_names()
-
-    def get_column_names_from_table(self, table_name: str) -> List[str]:
-        columns = self.get_columns_from_table(table_name)
-        names = [col.get("name") for col in columns]
-        return names
-
-    def get_columns_from_table(self, table_name: str) -> List[Dict]:
-        """Gets the columns form the given table
-
-        Args:
-            table_name (str): The id(name) of the table the columns will be returned from
-
-        Returns:
-            List[Dict]: A list of columns in dict form
-        """
-        inspector = sa.inspect(self.engine)
-        return inspector.get_columns(table_name)
-
-    def get_schemas(self) -> List[str]:
-        """Gets the database schemas
-
-        Returns:
-            List[str]: A list of names of the the schemas
-        """
-        inspector = sa.inspect(self.engine)
-        return inspector.get_schema_names()
-
-    def get_current_schema(self) -> str:
-        """Gets the current database schema
-
-        Returns:
-            str: The name of the current schema
-        """
-        return self.execute_sql_statement("SELECT DATABASE();").fetchall()[0][0]
-
-    def _change_current_schema(self, schema_name: str):
-        """Changes the current schema
-
-        Args:
-            schema_name (str): The name of the schema to change to
-        """
-        self.execute_sql_statement(f"USE {schema_name};")
-
-    def _create_schema(self, schema_name: str):
-        """Creates a schema
-
-        Args:
-            schema_name (str): The name of the schema to create
-        """
-        self.execute_sql_statement(f"CREATE DATABASE IF NOT EXISTS {schema_name};")
-
-    def _drop_schema(self, schema_name: str):
-        """Drops a schema
-
-        Args:
-            schema_name (str): The name of the schema to drop
-        """
-        self.execute_sql_statement(f"DROP DATABASE {schema_name};")
