+++ synapse/__init__.py
@@ -1,0 +1,3 @@
+""" Import all synapse classes
+"""
+from .synapse import *
